requirements.txt
requests
feedparser
python-dotenv
ruff
black
isort
pytest
pytest-mock
google-generativeai
python-telegram-bot
beautifulsoup4
